moves alls views to a subpackage
