+++ clashstats/views/__init__.py
@@ -1,0 +1,4 @@
+import clashstats.views.clan
+import clashstats.views.error
+import clashstats.views.index
+import clashstats.views.status
