+++ clashstats/views/error.py
@@ -1,0 +1,12 @@
+from clashstats import app
+from flask import render_template
+
+
+@app.errorhandler(404)
+def page_not_found(e):
+    return render_template('404.html'), 404
+
+
+@app.errorhandler(500)
+def page_not_found(e):
+    return render_template('500.html'), 500
